add o2 triplet test, clarify dsrg relax option 'twice'
